Fix isCellEditable by adding a mouse button as parameter
Fix for commit 7c14d67da625f741477351183633211a9164a69a

getModifiers() is deprecated and was previously replaced by
getModifiersEx(). Without the additional parameter the button was
internally first set to NOBUTTON but getModifiers() resulted in the
button being changed to BUTTON1. getModifiersEx(), on the other hand,
did not change the button.